mcp-servers/knowledge/requirements.txt
fastmcp>=0.1.0
langchain-pinecone>=0.2.0
langchain>=0.3.0
pinecone-client>=4.0.0
uvicorn[standard]>=0.30.0
